+++ src/index.d.ts
@@ -32,9 +32,3 @@ export { useAsyncRun } from './use-async-run';
 export { useAsyncCombineAll } from './use-async-combine-all';
 export { useAsyncCombineSeq } from './use-async-combine-seq';
 export { useAsyncCombineRace } from './use-async-combine-race';
-
-// custom async hooks
-
-export { useAsyncTaskTimeout } from './use-async-task-timeout';
-export { useAsyncTaskDelay } from './use-async-task-delay';
-export { useAsyncTaskFetch } from './use-async-task-fetch';