+++ src/index.js
@@ -4,7 +4,3 @@ export { useAsyncRun } from './use-async-run';
 export { useAsyncCombineAll } from './use-async-combine-all';
 export { useAsyncCombineSeq } from './use-async-combine-seq';
 export { useAsyncCombineRace } from './use-async-combine-race';
-
-export { useAsyncTaskTimeout } from './use-async-task-timeout';
-export { useAsyncTaskDelay } from './use-async-task-delay';
-export { useAsyncTaskFetch } from './use-async-task-fetch';