+++ src/use-async-task-delay.js
@@ -1,12 +1,12 @@
 import { useRef } from 'react';
-import shallowequal from 'shallowequal';
 
 import { useAsyncTaskTimeout } from './use-async-task-timeout';
+import { shallowArrayEqual } from './utils';
 
 export const useAsyncTaskDelay = (milliSeconds, inputs) => {
   const func = useRef();
   const prevInputs = useRef();
-  if (!prevInputs.current || !shallowequal(prevInputs.current, inputs)) {
+  if (!prevInputs.current || !shallowArrayEqual(prevInputs.current, inputs)) {
     prevInputs.current = inputs;
     func.current = () => true;
   }